Refactor parking spaces confirmations column
